requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
jalali_core==1.0.0
jdatetime==5.2.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
